feat: improve dmin dmax handling
